Add env file that works with `make brew`
To try and make it easier for new devs I've tried to simplify setting up
the environment for a dmrunner system on macOS.

I've added an file to source, `Brewfile.env`, that can be used to load
the required environment variables if `make brew` has been run. `make
brew` will then suggest you source that file and add it to your profile.