Allow DM runner not to manage nginx
On some systems, DM runner can't control nginx. For instance, in Linux with rootless docker, it doesn't have the necessary permissions.

So only record if nginx is not running. This way, dm runner can now cope with nginx being managed externally and still deal with the other background services itself.