Add environment specifier to dataclasses requirement
For versions of Python < 3.7, black requires the dataclasses
package[[1]]. However, for versions of Python >= 3.7 dataclasses is not
available on PyPI, so attempting to install `dataclasses==0.8` will
fail.

This commit tweaks the `requirements.txt` file by adding an environment
marker[[2]] such that for Python >= 3.7 `pip` will not try to install
the dataclasses package.

[1]: https://github.com/psf/black/blob/6d0bdc2f38d57ddf07e7d127215d51f8553997f9/Pipfile#L34
[2]: https://www.python.org/dev/peps/pep-0508/#environment-markers